third_party/blink/web_tests/external/wpt/css/css-lists/parsing/counter-reset-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['counter-reset'] = "reversed(chapter)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['counter-reset'] = "chapter reversed(chapter)" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['counter-reset'] = "reversed(chapter) 9 chapter" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['counter-reset'] = "reversed(section) -1" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['counter-reset'] = "reversed(first) -1 reversed(second) third 99" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
